Fix TestSpatialTree tests
Their use of grow is wrong for the new bbox. TestSpatialTree was expecting that the larger number that BBox was 'grown' to would not be in the set. That is not the case anymore. However their request operations can be had through the min() max() and Vector version of BBox's constructor.